pyproject.toml
[tool.poetry]
name = "landfiller"
version = "0.1.0"
description = "A tool to add landfill under entities in Factorio blueprints."
authors = ["Emil 'Skeen' Madsen <sovende@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
factorio-draftsman = "^1.1.1"
typer = {extras = ["all"], version = "^0.9.0"}
pyperclip = "^1.8.2"


[tool.poetry.group.pre-commit.dependencies]
ruff = "^0.1.15"
pre-commit = "^3.6.0"
reorder-python-imports = "^3.12.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
